+++ src/main/java/org/deepsymmetry/beatlink/data/ArtFinder.java
@@ -150,7 +150,7 @@ public class ArtFinder extends LifecycleParticipant {
      * @param announcement the packet which reported the device’s disappearance
      */
     private void clearArt(DeviceAnnouncement announcement) {
-        final int player = announcement.getNumber();
+        final int player = announcement.getDeviceNumber();
         // Iterate over a copy to avoid concurrent modification issues
         for (DeckReference deck : new HashSet<DeckReference>(hotCache.keySet())) {
             if (deck.player == player) {
@@ -281,7 +281,7 @@ public class ArtFinder extends LifecycleParticipant {
                                             final boolean failIfPassive) {
 
         // First check if we are using cached data for this slot.
-        MetadataCache cache = MetadataFinder.getInstance().getMetadataCache(SlotReference.getSlotReference(artReference));
+        @SuppressWarnings("deprecation") MetadataCache cache = MetadataFinder.getInstance().getMetadataCache(SlotReference.getSlotReference(artReference));
         if (cache != null) {
             final AlbumArt result = cache.getAlbumArt(null, artReference);
             if (result != null) {