+++ src/main/java/org/deepsymmetry/beatlink/data/WaveformPreviewComponent.java
@@ -535,6 +535,7 @@ public class WaveformPreviewComponent extends JComponent {
                         public void run() {
                             while (animating.get()) {
                                 try {
+                                    //noinspection BusyWait
                                     Thread.sleep(33);  // Animate at 30 fps
                                 } catch (InterruptedException e) {
                                     logger.warn("Waveform animation thread interrupted; ending");